refactor: Id long 타입으로 변경

+++ Assets/Scripts/Data/ClientContext.cs
@@ -1,13 +1,13 @@
 
 public static class ClientContext
 {
-    public static int PlayerId { get; private set; }
+    public static long PlayerId { get; private set; }
 
     public static string Nickname { get; private set; }
 
 
     public static void SetLogin(
-        int playerId,
+        long playerId,
         string nickname)
     {
         PlayerId = playerId;
